Moved the logo
